+++ src/app/src/components/Dashboard.jsx
@@ -14,6 +14,8 @@ import DashboardUpdateFacilityLocation from './DashboardUpdateFacilityLocation';
 import DashboardApiBlocks from './DashboardApiBlocks';
 import DashboardActivityReports from './DashboardActivityReports';
 import DashboardApiBlock from './DashboardApiBlock';
+import DashboardLinkToOarId from './DashboardLinkToOarId';
+import DashboardGeocoder from './DashboardGeocoder';
 import FeatureFlag from './FeatureFlag';
 import RouteNotFound from './RouteNotFound';
 
@@ -32,6 +34,8 @@ import {
     dashboardApiBlocksRoute,
     dashboardApiBlockRoute,
     dashboardActivityReportsRoute,
+    dashboardGeocoderRoute,
+    dashboardLinkOarIdRoute,
 } from '../util/constants';
 
 import AppGrid from './AppGrid';
@@ -87,16 +91,18 @@ function Dashboard({ userWithAccessHasSignedIn, fetchingSessionSignIn }) {
             <FeatureFlag flag={CLAIM_A_FACILITY}>
                 <Link to={dashboardClaimsRoute}>View Facility Claims</Link>
             </FeatureFlag>
-            <Link to={dashboardDeleteFacilityRoute}>Delete a facility</Link>
-            <Link to={dashboardMergeFacilitiesRoute}>Merge two facilities</Link>
+            <Link to={dashboardDeleteFacilityRoute}>Delete a Facility</Link>
+            <Link to={dashboardMergeFacilitiesRoute}>Merge Two Facilities</Link>
             <Link to={dashboardAdjustFacilityMatchesRoute}>
-                Adjust facility matches
+                Adjust Facility Matches
             </Link>
             <Link to={dashboardUpdateFacilityLocationRoute}>
-                Update facility location
+                Update Facility Location
             </Link>
             <Link to={dashboardApiBlocksRoute}>View API Blocks</Link>
             <Link to={dashboardActivityReportsRoute}>View Status Reports</Link>
+            <Link to={dashboardLinkOarIdRoute}>Link to New OAR ID</Link>
+            <Link to={dashboardGeocoderRoute}>Geocode</Link>
         </div>
     );
 
@@ -192,6 +198,18 @@ function Dashboard({ userWithAccessHasSignedIn, fetchingSessionSignIn }) {
                         />
                         <Route
                             exact
+                            path={dashboardLinkOarIdRoute}
+                            render={makeClickableDashboardLinkFn(
+                                'Link to New OAR ID',
+                            )}
+                        />
+                        <Route
+                            exact
+                            path={dashboardGeocoderRoute}
+                            render={makeClickableDashboardLinkFn('Geocoder')}
+                        />
+                        <Route
+                            exact
                             path={dashboardRoute}
                             render={() => 'Dashboard'}
                         />
@@ -262,6 +280,16 @@ function Dashboard({ userWithAccessHasSignedIn, fetchingSessionSignIn }) {
                         exact
                         path={dashboardActivityReportsRoute}
                         component={DashboardActivityReports}
+                    />
+                    <Route
+                        exact
+                        path={dashboardLinkOarIdRoute}
+                        component={DashboardLinkToOarId}
+                    />
+                    <Route
+                        exact
+                        path={dashboardGeocoderRoute}
+                        component={DashboardGeocoder}
                     />
                     <Route
                         exact