+++ src/app/src/components/FacilityDetailSidebar.jsx
@@ -1,4 +1,5 @@
 import React, { Component } from 'react';
+import { Redirect } from 'react-router';
 import { connect } from 'react-redux';
 import { arrayOf, bool, func, shape, string } from 'prop-types';
 import CircularProgress from '@material-ui/core/CircularProgress';
@@ -156,6 +157,13 @@ class FacilityDetailSidebar extends Component {
             classes,
         } = this.props;
 
+        if (data?.id && data?.id !== oarID) {
+            // When redirecting to a facility alias from a deleted facility,
+            // the OAR ID in the url will not match the facility data id;
+            // redirect to the appropriate facility URL.
+            return <Redirect to={`/facilities/${data.id}`} />;
+        }
+
         if (fetching) {
             return (
                 <div className="control-panel">
@@ -238,6 +246,7 @@ class FacilityDetailSidebar extends Component {
         const facilityClaimID = get(data, 'properties.claim_info.id', null);
 
         const report = get(data, 'properties.activity_reports[0]');
+        const newOarId = get(data, 'properties.new_oar_id');
         const renderStatusRibbon = () => {
             if (!report) return null;
             if (report.status === 'PENDING') {
@@ -254,7 +263,17 @@ class FacilityDetailSidebar extends Component {
                 return (
                     <FeatureFlag flag={REPORT_A_FACILITY}>
                         <div className={classes.closureRibbon}>
-                            This facility is closed.
+                            This facility is closed
+                            {!!newOarId && (
+                                <span>
+                                    {' '}
+                                    and has moved to{' '}
+                                    <a href={`/facilities/${newOarId}`}>
+                                        {newOarId}
+                                    </a>
+                                </span>
+                            )}
+                            .
                         </div>
                     </FeatureFlag>
                 );