+++ src/django/api/models.py
@@ -1288,6 +1288,12 @@ class Facility(PPEMixin):
         null=True,
         help_text=('Whether this facility is closed.')
     )
+    new_oar_id = models.CharField(
+        max_length=32,
+        null=True,
+        blank=True,
+        help_text=('The new OAR ID where this facility can be found if it has '
+                   'been moved.'))
     created_at = models.DateTimeField(auto_now_add=True)
     updated_at = models.DateTimeField(auto_now=True, db_index=True)